+++ backend/src/api/routes/configs.ts
@@ -1,6 +1,6 @@
 import { Router } from "express";
 import { authenticateRequest } from "../../middlewares/auth";
-import { asyncHandler, validateRequest } from "../../middlewares/api-utils";
+import { asyncHandler, validateRequest } from "../../middlewares/utility";
 import configSchema from "../schemas/config-schema";
 import * as ConfigController from "../controllers/config";
 import * as RateLimit from "../../middlewares/rate-limit";
