+++ backend/src/api/routes/webhooks.ts
@@ -1,7 +1,7 @@
 // import joi from "joi";
 import { Router } from "express";
 import { authenticateGithubWebhook } from "../../middlewares/auth";
-import { asyncHandler } from "../../middlewares/api-utils";
+import { asyncHandler } from "../../middlewares/utility";
 import { webhookLimit } from "../../middlewares/rate-limit";
 import { githubRelease } from "../controllers/webhooks";
